Split legacy capture support helpers
